requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.5
cloudinary==1.25.0
cryptography==3.4.7
Django==3.2.4
django-bootstrap4==3.0.1
django-cors-headers==3.7.0
djangorestframework==3.12.4
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.1.0
python-decouple==3.4
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.5
